+++ OptiHub/Models/HubSection.cs
@@ -2,6 +2,7 @@ namespace OptiHub.Models;
 
 public enum HubSection
 {
+    Home,
     Apps,
     Internet,
     Gpu